requirements.txt
git+https://github.com/openai/whisper.git
googletrans==4.0.0rc1
torch>=1.13.1
youtube_dl>=2021.12.17certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
colorama==0.4.6
filelock==3.12.4
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.24.4
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
pytube==15.0.0
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
youtube-dl==2021.12.17
zipp==3.17.0
